requirements.txt
flask==2.0.1
Pillow==10.0.0
torch==2.0.1
torchvision==0.15.2
timm==0.9.2
PyMuPDF==1.22.5
Werkzeug==2.0.1 